requirements.txt
python-telegram-bot==13.15
requests
beautifulsoup4
gspread
google-auth-oauthlib
selenium
webdriver-manager
google-generativeai
fpdf2